% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_defaults.R
\name{add_defaults}
\alias{add_defaults}
\title{Adds default styles to table}
\usage{
add_defaults(table)
}
\arguments{
\item{table}{The table to add the styles to}
}
\value{
Modified SimplePlotTable with data and styles
}
\description{
Adds default styles to table
}
