% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot.SimplePlotTable}
\alias{autoplot.SimplePlotTable}
\title{Plot a SimplePlotTable}
\usage{
\method{autoplot}{SimplePlotTable}(object, ...)
}
\arguments{
\item{object}{The SimplePlotTable to plot}

\item{...}{Required to override \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}}, ignored for now.}
}
\value{
A ggplot object containing the table
}
\description{
Creates a plot from the given SimplePlotTable using void theme.
You can hide headers if desired.
}
\details{
You can apply all functions to the result that you would use if ordinary plots from ggplot.
The column and row headers are drawn as labels on the x and y axis, respectively.
You can change their appearance using \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\examples{
table <- new_SimplePlotTable(mtcars)
autoplot(table)
}
