% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_geom}
\alias{create_geom}
\title{Create a table's text geom}
\usage{
create_geom(table)
}
\arguments{
\item{table}{The table object}
}
\value{
A ggplot text geom representing the data
}
\description{
Create a table's text geom
}
