% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotTable.R
\name{is.SimplePlotTable}
\alias{is.SimplePlotTable}
\title{Checks if the argument is a SimplePlotTable}
\usage{
is.SimplePlotTable(table)
}
\arguments{
\item{table}{The argument to check}
}
\value{
\code{TRUE} if the argument is a SimplePlotTable
}
\description{
Checks if the argument is a SimplePlotTable
}
