% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotTable.R
\name{new_SimplePlotTable}
\alias{new_SimplePlotTable}
\title{Construct a new SimplePlotTable}
\usage{
new_SimplePlotTable(
  data,
  col_names = colnames(data),
  row_names = rownames(data)
)
}
\arguments{
\item{data}{The data.frame that should be displayed}

\item{col_names}{The column headers to display, defaults to data's colnames}

\item{row_names}{The row headers to display, defaults to data's rownames}
}
\value{
A new SimplePlotTaböe
}
\description{
Construct a new SimplePlotTable
}
