% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Converts a data.frame to long format with indexes}
\usage{
prepare_data(data)
}
\arguments{
\item{data}{The data to transform}
}
\value{
The transformed data, with columns row, col and text
}
\description{
Converts a data.frame to long format with indexes
}
