% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_style.R
\name{set_column_header_style}
\alias{set_column_header_style}
\title{Set the appearance of the column headers}
\usage{
set_column_header_style(table, ...)
}
\arguments{
\item{table}{The table to modify}

\item{...}{Styling options for the text, passed to \code{\link[ggplot2:element]{ggplot2::element_text()}}}
}
\value{
The modified table
}
\description{
Set the appearance of the column headers
}
\examples{
# Print column headers in bold
table <- set_column_header_style(table, face="bold")

# Set headers to flush right
table <- set_column_header-style(table, hjust=1)
}
