% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_style.R
\name{set_style}
\alias{set_style}
\title{Change the appearance of cells in a table}
\usage{
set_style(table, rows = 1:length(table$rows), cols = 1:length(table$cols), ...)
}
\arguments{
\item{table}{The SimplePlotTable to change}

\item{rows}{The indexes of rows which should be styled, defaults to all rows}

\item{cols}{The indexes of columns which should be styled, defaults to all columns}

\item{...}{The properties to change. Argument names must be aesthetics accepted by \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}
}
\value{
A SimplePlotTable with changed styling.
}
\description{
Allows you to set the visual appearance of single cells or ranges in a table
}
\details{
This function allows for styling of cells in a SimplePlotTable.
The cells to be changed are identified by \code{rows} and \code{cols}.
Every cell that is in any of the rows and any of the columns specified is changed.
}
\examples{
# Make all cells blue
table <- set_style(table, col="blue")

# Make the second column italic
table <- set_style(table, cols=2, fontface="italic")

# Make rows 2 through 4 diagonal and typeset in Times New Roman
table <- set_style(table, rows=2:4, family="Times New Roman", angle=45)

# Make the range of cells spanning from (2,2) to (5,8) bold
table <- set_style(table, rows=2:8, cols=2:5, fontface="bold")
}
