% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot.SimplePlotTable}
\alias{autoplot.SimplePlotTable}
\title{Plot a SimplePlotTable}
\usage{
\method{autoplot}{SimplePlotTable}(table, col_headers = TRUE, row_headers = TRUE)
}
\arguments{
\item{table}{The SimplePlotTable to plot}

\item{col_headers}{Logical indicating whether column headers should be shown (default: TRUE)}

\item{row_headers}{Logical indicating whether row headers should be shown (default: TRUE)}
}
\value{
A ggplot object containing the table
}
\description{
Creates a plot from the given SimplePlotTable using void theme.
You can hide headers if desired.
}
\details{
You can apply all functions to the result that you would use if ordinary plots from ggplot.
The column and row headers are drawn as labels on the x and y axis, respectively.
You can change their appearance using \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\examples{
# Plot with all headers
create_plot(table)

# Plot column headers only
create_plot(table, T, F)

# Plot without headers
create_plot(table, F, F)
}
